requirements.txt
aws-cdk-lib==2.177.0
constructs>=10.0.0,<11.0.0
cdk-aws-lambda-powertools-layer
aws-lambda-powertools
